разделил векторы и матрицы
